Secure cookie auth & SignalR real-time notifications
- Implement HttpOnly, Secure cookie-based JWT auth (login, refresh, logout)
- Add logout and "me" endpoints to AuthController
- Extract JWT from cookies in auth middleware
- Expose Set-Cookie header in CORS for credentials
- Add SignalRController with negotiate endpoint for short-lived tokens
- Introduce IRealTimeNotifier and SignalRNotifier for chat events
- Update ChatHub for online/offline, typing, delivery/read events
- MessagingService now triggers real-time notifications
- Add/revise tests for logout and real-time messaging
- Refactor for nullability, logging, and DI registration

+++ PutZige.Application.Tests/TestJwtTokenService.cs
@@ -5,7 +5,7 @@ namespace PutZige.Application.Tests
 {
     internal class TestJwtTokenService : IJwtTokenService
     {
-        public string GenerateAccessToken(Guid userId, string email, string username, int expiryMinutes, out DateTime expiresAt)
+        public string GenerateAccessToken(Guid userId, string email, string username, double expiryMinutes, out DateTime expiresAt)
         {
             expiresAt = DateTime.UtcNow.AddMinutes(expiryMinutes);
             return "access-token-test";
