+++ PutZige.Application/Interfaces/IJwtTokenService.cs
@@ -6,7 +6,7 @@ namespace PutZige.Application.Interfaces
 {
     public interface IJwtTokenService
     {
-        string GenerateAccessToken(Guid userId, string email, string username, int expiryMinutes, out DateTime expiresAt);
+        string GenerateAccessToken(Guid userId, string email, string username, double expiryMinutes, out DateTime expiresAt);
         string GenerateRefreshToken();
         ClaimsPrincipal? ValidateToken(string token, bool validateLifetime = true);
     }
